temp directory for completion tests

+++ util/temp_dir.py
@@ -7,17 +7,25 @@ class ScopedTempDirectory(object):
     self._old_directory = None
     self._delete_on_exit = not temp_directory
     self._delete_non_empty = delete_non_empty
+    self._exited = True
 
   def __enter__(self):
+    self._exited = False
     if not self._temp_directory:
       self._temp_directory = tempfile.mkdtemp()
     self._old_directory = os.getcwd()
     os.chdir(self._temp_directory)
 
   def __exit__(self, *args):
+    if self._exited:
+      return
+    self._exited = True
     os.chdir(self._old_directory)
     if self._delete_on_exit:
       if self._delete_non_empty:
         os.system('rm -rf {}'.format(self._temp_directory))
       else:
         os.rmdir(self._temp_directory)
+
+  def __del__(self):
+    self.__exit__()
